refactor/docs: change compressor name & add docs
