+++ lib/src/auto_cache_injections.dart
@@ -3,7 +3,7 @@ import 'package:auto_cache_manager/auto_cache_manager.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'core/core.dart';
-import 'core/services/compresser/compresser_service.dart';
+import 'core/services/compressor/compressor_service.dart';
 import 'core/services/cryptography/encrypt/encrypt_cryptography_service.dart';
 import 'core/services/cryptography/i_cryptography_service.dart';
 import 'core/services/directory_provider/path_provider/path_provider_service.dart';
@@ -26,7 +26,7 @@ class AutoCacheInjections {
 
     Injector.I.bindSingleton<CacheConfig>(AutoCacheManagerInitializer.I.config);
     Injector.I.bindSingleton<IPathProviderService>(PathProviderService());
-    Injector.I.bindSingleton<ICompresserService>(CompresserService());
+    Injector.I.bindSingleton<ICompressorService>(CompressorService());
 
     Injector.I.bindSingleton<IPrefsService>(SharedPreferencesService(Injector.I.get<SharedPreferences>()));
